refactor(components): refactor result
