+++ packages/components/result/__tests__/result.spec.ts
@@ -1,5 +1,5 @@
 import makeMount from '@element-plus/test-utils/make-mount'
-import Result from '../src/index.vue'
+import Result from '../src/result.vue'
 
 const AXIOM = 'Rem is the best girl'
